# Use ThreadStatic for cached fields