--- a/src/GraphQL/Execution/ExecutionContext.cs
+++ b/src/GraphQL/Execution/ExecutionContext.cs
@@ -169,20 +169,4 @@ public class ExecutionContext : IExecutionContext, IExecutionArrayPool, IDisposa
             _trackedArrays.Clear();
         }
     }
-
-    /// <summary>
-    /// Allows for an execution strategy to reuse an instance of <see cref="ReadonlyResolveFieldContext"/>.
-    /// This field may be accessed by multiple threads at the same time, so
-    /// access is restricted to <see cref="Interlocked.Exchange{T}(ref T, T)"/>
-    /// and <see cref="Interlocked.CompareExchange{T}(ref T, T, T)"/>.
-    /// </summary>
-    internal ReadonlyResolveFieldContext? ReusableReadonlyResolveFieldContext;
-
-    /// <summary>
-    /// Allows for an execution strategy to reuse an instance of <see cref="Dictionary{TKey, TValue}"/>.
-    /// This field may be accessed by multiple threads at the same time, so
-    /// access is restricted to <see cref="Interlocked.Exchange{T}(ref T, T)"/>
-    /// and <see cref="Interlocked.CompareExchange{T}(ref T, T, T)"/>.
-    /// </summary>
-    internal Dictionary<string, (GraphQLField field, FieldType fieldType)>? ReusableFields;
 }